restructure: Fix import paths
