+++ src/UI/containers/Panels/assemblies/index.js
@@ -1,0 +1,2 @@
+export {default as ArchivePanel} from './ArchivePanel'
+export {default as WatchPanel} from './WatchPanel'
