+++ src/UI/containers/Panels/components/Panel/Panel.styles.sass
