+++ src/UI/containers/Panels/components/Panel/index.js
