+++ src/UI/containers/index.js
@@ -1,8 +1,8 @@
 export {default as App} from './App';
 export {default as Board} from './Board';
-export {default as ContentPanels} from './ContentPanels';
 export {default as Dashboard} from './Dashboard';
 export {default as Header} from './Header';
 export {default as MediaViewer} from './MediaViewer';
+export {default as Panels} from './Panels';
 export {default as Settings} from './Settings';
 export {default as Thread} from './Thread';
