+++ src/UI/index.jsx
@@ -10,6 +10,6 @@ import '-/config/client.settings';
 
 import React from 'react';
 import ReactDOM from 'react-dom';
-import App from './modules/App';
+import App from './containers/App';
 
 ReactDOM.render(<App />, document.querySelector('#App'));
