+++ src/UI/views/MediaView.jsx
@@ -1,7 +1,7 @@
 import React, { PropTypes } from 'react';
 import cx from 'classnames'
 
-import MediaViewer from '~/modules/MediaViewer';
+import MediaViewer from '~/containers/MediaViewer';
 
 const MediaView = ({ className }) => {
     return (
